Add initial support for Identity-V encoding.
This is a simple fix that enables the access to texts in Identity-V
encoding.  Identity-V encoding is vertical version of Identity-H and the
mapping is the same as for Identity-H (See Table 118 in PDF
32000-1:2008).

This change dosen't distinguish between Identity-H and Identity-V
(because thier mapping is the same).
But it would be better to support to know whether the targetd text is in
horizontal layout or in vertical layout)