generate a path by which to send layout engine options
I am not sure that there is anyway to always predict the best column
width for any given pdf, to that end I created a way to pass a fudge
factor off to the layout engine.
Additionally I added a check to the layout engine that emits a warning
on stderr if a problem might occur.