Fixed character spacing Tc so that it works properly
Tc (and Tw) wasn't implemented in a way that generated correct results.
Added a bunch of notes (in code) and created a spec for testing
results. The pdf file has lines in it spaced at 0pt, 144 pt, 288 pt,
and one other place. the width of the wide character spaced string
should be about 277.