Fix inline image parsing: EI does not need prefixed whitespace
I came across an interesting example of a PDF from my bank statement.
An embedded image in the PDF looks like this:

BI\nID \x00\x00\x00\x00EI

As far as I can tell, there's no detail in the spec that says an
embedded image must be terminated with a whitespace character. Since
this PDF is parse-able, I think a simple fix can allow for this corner
case while sill allowing other examples where the "EI" string correctly
appears within the image (not a terminating token).

Due to the null character being appended (and a new corner case to make
sure this char is appended), I removed the "strip" call on the string
token. I traced through the commit history, but could not find a good
reason for why this "strip" call was added. Removing this did not cause
any specs to fail (which is not a great heuristic, but at least show
this might not actually be important). If it is important, then we can
at least parse the text correctly (which is good enough for me).

Resolves issue #169