Fix decryption of some PDFs
PDFs encrypted with revision 4 (or higher) of the standard handler and
with metadata unencrypted are required to pad in some extra bytes at a
specific step.

We had two mistakes that prevented it from working:

- we were looking for the wrong revision
- we'd misspelt an instance variable

There's now an integration spec that checks this type of file can be
decrypted, so we should be safe from future regressions.

Thanks to Gyuchang Jun for originally pointing this out [1].

[1] https://github.com/yob/pdf-reader/pull/161