add a sample PDF that has a form xobject that references itself
This single page PDF has a Form XObject in its Resources dict. The Form
XObject has its own Resources dict, which has an XObject pointing to
itself.

Nested XObjects are fine, but not recursive ones.

This is a simpler file that demonstrates an issue originally found in
the PDF provided at [1]

[1] https://github.com/yob/pdf-reader/issues/94#issuecomment-272663109