refactor CMap parsing to improve compatability with valid CMap's
We previously assumed that CMaps would include line breaks in specific
places, which turns out to be true in *most* cases, but not all.

It turns out CMaps follow the same basic syntax as other parts of PDF
files (thanks to the postscript heritage) and we can leverage our
standard parser to handle CMaps that are less conventional (but
perfectly valid) in their use of line endings.