add a pending integration spec for a file we could handle better
Most type3 fonts have a FontMatrix entry of [ 0.001 0 0 0.001 0 0 ],
which matches the glyph scale factor of 1000 that non-type3 fonts use.
It's permitted for type3 fonts to use other FontMatrix values though,
and we should do a better job of extracting the text.

Thanks to pdf.js for this sample file, it was originally
(test/pdfs/issue5896.pdf) in their repo, and is a simpler test case of a
private PDF I was sent.