add some failing specs for encrypted files we don't yet support
We already had some sample files generated with Adbobe Acrobat, but had
no implemented specs to confirm how we handled them.

Turns out we raise cryptic errors that don't make it obvious to the user
that we don't support some types of encryption yet. I'll address that in
the next commit.