replicate two integration specs into unit specs
In the previous commit I added some integration specs that confirm our
behaviour when extracting text from PDFs with overlapping characters as
a fake "bold" effect.

The fix is likely to be in PDF::Reader::PageLayout, so this takes a
small part of the data from those PDFs and creates a unit spec that can
help drive out a solution.