Fallback to a substitute font for Type1 and TrueType fonts that aren't embedded
Some PDFs use fonts that aren't embedded and are either:

* Type1, yet aren't one of the 14 "built in" fonts that PDF readers are
  expected to support
* TrueType, and PDF readers will attempt to find a font with the same
  name on the local system. If there's no match, an
  approximate substitute will be used.

In these situations we've raised an exception. That's frustrating for
users and not what other readers do, so this changes pdf-reader to
perform a font substitution.

For now the substitution is super-simple. We don't look for a matching
font on the local system, we just use the font metrics for the built-in
Times-Roman font.

That might be enough for now. pdf-reader isn't attempting to render any
pages visually, so we only need the substitute for glyph metrics (like
width).

Closes #292
Closes #307