convert from runtime sigs to storing them in an rbi/ directory
Shipping the RBI file with the gem like this means downstream users of
pdf-reader who also use sorbet will know the types pdf-reader expects
for its public API.

Storing the sigs outside the source files also means there's no need to
add sorbet-runtime as a dependency of the gem, and downstream users of
pdf-reader who do not use sorbet will see no change.

I generated the RBI file with parlour:

    bundle exec parlour

I assume I'll have to regenerate it over time as the methods on each
class evolve. That sounds like a huge hassle, but this is an experiment
so lets see how it goes.

The parlour config file ignore lib/pdf/reader/parser.rb because I get a
NoMethodError if I leave it in.