Fix glyph width calculations for some Type3 fonts
Most type3 fonts have a FontMatrix entry of [ 0.001 0 0 0.001 0 0 ],
which matches the glyph scale factor of 1000 that non-type3 fonts use.
It's permitted for type3 fonts to use other FontMatrix values though,
and now we use the FontMatrix for glyph width calculations in type3
fonts.

A few integration specs for non-type3 PDFs needed tweaking thanks to the
the font size in the TextRun no longer being floored. The changes are
whitespace only though, and don't seem to be significant.

Fixes the pending spec added #200