Avoid providing a initial value for instance var type hints in RBI file
These type hints were added in #415 while changing some files to `typed:
strict`. It felt weird providing an initial value that I knew wasn't
used, and in a follow up comment on the PR it was suggested that I can
use this form to be clear that I'm annotating the ivar type without
setting a value.

`T.unsafe(nil)` also feels a bit weird. So much syntax for so little
value. It'd be nice if sorbet had a way to only specificy the ivar type
in an RBI files, without the value argument at all 🤷‍♂️