Add type checking derefs to ObjectHash and use where we can
The previous commit adds a bunch of corrupt PDF files with specs to
ensure they raise appropriate errors.

The bulk of the issues are incorrect types coming out of the PDF, and
pdf-reader not noticing. For example, we expect an Array but get back a
Number and then start passing it around until we eventually call .each
on it and get a NoMethodError.

If the PDF is corrupt and has an unusable object type, I think we want to
raise a MalformedPDFError as close to the source of the problem as
possible. This commit proposes doing that by add type-specific deref
methods to ObjectHash. When using these new methods we are asserting the
type we expect back, and for valid PDFs there's no change in behaviour.
However, PDFs that have the wrong type for a particular object (and the
type can't be cast/coerced into the required type) will raise a useful
error.

There's still a lot more we could do for type safety. In particular,
reading values out of Hashes/Dicts is still not type safe in many
places. Still, this is a good start that will pick up many basic
corruption issues.