Fix endless hang when processing PDF with invalid xref table
This PDF has an xref table that starts with the token `<` instead of a
numeric object ID. We force it to an int of 0, but byu then we've
interpreted the following token as a hex string and we think the xref
table has 85,000,000,000,065,535 entries.

I considered adding a cap on the number of xref table entries we'll
process (1 million? something implausibly high) and raising
MalformedPDFError when that cap is exceeded.

Maybe we want that as well, but I think in this case I want to fail
early and raise an error that targets the root cause of the issue.

I'm not thrilled with the complexity of this method in the XRef class,
and this new conditional doesn't make it any better. I wanted to focus
on adding a test case and fixing it though. I'll try to refactor the
class separately.

Fixes #217