Enable type checking of the CMap class
This was originally skipped in #361 because I was getting an unreachable
code error. I found it was possible to avoid that by providing a type
annotation for the `mode` variable of the state machine instead of hard
coding it to `:none`.

Sadly, I can't do the type annotation with `T.let()` because sorbet
isn't a runtime dependency of pdf-reader, but I can hack it in via a sig
in the RBI file if I make the initial state part of the method params.