Make an effort to compensate when an xref offset is a few bytes too low
A sample file provided in #213 has an xref table where at least one of
the byte offsets is a few bytes too low. When we jump to that offset, we
don't find the start of the object as we expect.

However, the following token in the file is the start of the expected
object. Other readers I tested with seem to handle this particular
corruption gracefully, and now we do too.

Fixes #213