PDF::Reader::ObjectHash#page_references should return an array of PDF::Reader::Reference
In #421 I refactored PDF::Reader::ObjectHash to check they types of
objects in pulls from the PDF file and raise when the object type is the
wrong type.

In doing so, I inadvertently changed the page_references method to
return an Array of PDF::Reader::Page objects, rather than the indirect
references to them.

For most users it won't matter, but it might for some so it's a bug.

Interestingly, the RBI file already claims this method must return an
array of PDF::Reader::Reference objects and the sorbet static type
checks weren't failing.

Fixes #443