add some nil safety to LZW decompression
Based on the same data provided in #449.

For now I think I'd prefer to err on the side of strictness and raise
one of our documented errors when a PDF file contains invalid LZW
compressed data.

I'm mulling over how to handle this sort of issue long term though. It's
definitely true that other PDF readers tend to treat this kind of thing
as a soft error and do their best to render the rest of the page anyway.