change PDF::Reader to `typed: strict`
slowly slowly rolling out strict.

This was a fun one - sorbet pointed out that the type check in root()
isn't needed any more - type typed deref_hash() does that for us.

Also the comment and type hint for the metadata method was wrong. The
method returns a String, not a Hash