Ensure PDFs can be read from Tempfile instances
In #490 I tightened up the types of IO-like objects that pdf-reader will
accept, to help sorbet detect type errors. Previously we accepted a duck
typed object with {read,seek}, but from that PR on we accepted a file
path, an IO, or a StringIO.

However, I missed the common use case of passing a Tempfile instance.
Tempfile is in the stdlib, but turns out it's not an IO. It's most
commonly used for this:

    PDF::Reader.open(URI.open("https://example.com/foo.pdf")) do |pdf
      puts pdf.page_count
    end