Add advanced text run filter for text extraction
Some special PDFs (e.g. watermark as text in the background)
seem to break the text rendering of PageLayout, because it relies
on the mean font size and median glyph width. Which could lead to
silently overwriting text which should belong to another row.

In order to prevent this we add an option to select which text
should be extracted. With that you can exclude special text, that
is not relevant for the parsing.