re-enable strict typing in two files
After a recent upgrade to sorbet (#546) I found that I was able to
re-enable strict typing here.

It was disabled in #513 because sorbet wasn't finding the type
annotation in the rbi file for some constants. It's still inconsisent
(for example, I wasn't able to change the cmap.rb back to strict), but
these ones are OK.