Use sourcegraph amp to start converting from external rbi to inline rbs
Shopify recently announced support for defining types using RBS-style syntax
as inline comments [1]. This was appealing to my use case in a gem where:

* I don't want a runtime dependency on sorbet, so can't use the classic sorbet
  `sig {}` syntax
* When working on pdf-reader I'd like to have an LSP server assist me with
  method signatures and code changes
* I want to publish an RBI file in the gem so users of pdf-reader can type
  check their calls to classes in the PDF::Reaer namespace

It also felt like a good excuse to try amp [3] for a mechanical refactor. The
experience was mixed: I spent ~USD$20 down a few bad prompt threads, before
eventually landing on these changes. It's a step in the direction, but there's
lots of lcean up to do in a follow up commit.

[1] https://railsatscale.com/2025-04-23-rbs-support-for-sorbet/
[2] https://yob.id.au/2021/12/31/sorbet-with-gems.html
[3] https://ampcode.com/manual