remove parlour gem from Gemfile
When I first introduced sorbet to this project I used parlour to generate
an RBI file and work with it. However, now that I'm switching to inline
RBS comments for types, I don't think I need it any more.

The `spoom` gem is now my go-to for generating an RBI file for inclusion
in a gem release