dynamically generate an RBI file for inclusion in the gem
Now that we're using rbs style comments to define types, we don't have an
RBI file lying around to include in the gem. I think we still *want* one
in the gem though, so pdf-reader users can type check the way they use
the gem and so their LSP using editors can be extra helpful.

We can use `spoom` to dynamically generate the RBI file from the comments.