Add specs for two Security Handlers:
* PDF::Reader::AesV2SecurityHandler
* PDF::Reader::AesV3SecurityHandler

I'm planning to work on fixing the bug reported in #534, but first I want
the current behaviour speced. I did have to introduce a small behaviour change,
to validate the key length in AESV3SecurityHandler. That's because when
the key length is invalid we get different errors depending on the version
of ruby used, and we run tests on ruby 2.1+. Validating the key length
and raising the same error on all rubies felt like the best option.

I used amp [1] to help construct these, and it saved a buncha time

[1] https://ampcode.com