Handle PDFs that use AES encryption with no padding
From what I can tell this is a non-standard way to use AES encryption, but
of course there are PDFs in the wild that do it and other PDF readers seem
to handle it.

The two files changing behaviour here are used for:

* encryption v4, revision 4 with 128bit AES encryption
* encryption v5, revsions 5 with 256bit AES encryption

The spec changes here were largely written by amp [1], and the changes to get them passing
by me.

Fixes #534

[1] https://ampcode.com/