Avoid infinite recursion when a PDF has a loop in the ancestorsof a Page object
Each page in a PDF is a single Object in a tree of Pages->Pages->Page objects, and
the page inherits properties from its ancestors right up the tree.

We had not protection against a malicious (or incompetently generated) PDF
that built an ancestor tree with a cyclic loop. As a result, attempting to
parse such a PDF would raise a SystemStackError when the recursion hit a
max depth.

To fix it, when walking up the tree to find all the ancestors keep a record
of the ones we have seen and raise an error if we see one again.

Thanks to Yann @ Riema Labs for finding and reporting the issue.