Move back to an official release of spoom
It crashes if I have type definitions above the class like this, but works
if I move it to be inside the class.

Sorta.

It no longer crashes, but the exported signatures for methods using these
types in the sig come out as `untyped`. I can live with that for now, these
types are definitely working with sorbet itself and ruby-lsp is detecting
them, it's just an RBI export issue