Solve non-stream ToUnicode in a slightly different way
I've added a test case for the previous commit - a PDF where ToUnicode is
a PDF Name rather than a Stream containing a CMap.

With that in place, I decided to fixc it in a slightly different way - attempting
to dereference the value, and only then silently skipping loading ToUnicode
if it's not the expected type.

This means that we'll also catch the case where ToUnicode is an indirect
object that points to a non-stream