Skip invalid UTF-16 surrogate pairs in CMaps
CMaps used as a ToUnicode mapping should only return valid Unicode codepoints,
and the range 0xD800 -> 0xD8FF are not valid to return on their own. They
indicate the start of a surrogate pair and should be followed by more hex chars.

When they're not, lets just skip them from the mapping.

An existing spec had to be changed because it was asserting that an invalid
surrogate pair was being mapped.