Use Point class rather than [x,y] Arrays in more places
As I roll type hints out more broadly in this codebase, using `Array[Numeric]` to
represent an (x,y) point is a hassle. When reading from an array like [0,1], the standard
Array accessor methods can return a nil which makes sorbet unhappy and we need to
start adding guards to let sorbet know that the Array always has exactly two elements.

By switching to the Point class, sorbet can do more work for us. IF we have a Point
instance, both sorbet and we know that it *must* have a numeric x and y value available.

These two methods are part of the public API of pdf-reader so I was unable to change
their signature:

* PageState#ctm_transform
* PageState#ttm_transform

Instead, I've soft deprecated them and added these alternatives:

* PageState#ctm_transform_point
* PageState#ttm_transform_point